Replace documentation to the new one - New theme and new example for field to field function
