--- doc/_static/sphxglr.css
@@ -1,6 +1,0 @@
-/* hide intersphinx links from code cells
- * xref https://sphinx-gallery.github.io/stable/configuration.html#stylizing-code-links-using-css */
-a[class^="sphx-glr-backref-module-"] {
-    text-decoration: none;
-    color: inherit;
-}
