--- doc/application/application.rst
@@ -1,16 +1,0 @@
-Applications
-============
-
-otfmi bridges the physics of the model (in `FMI <https://fmi-standard.org/>`__ format) with the universe of
-`OpenTURNS <http://openturns.github.io/openturns/master/contents.html>`__ probabilistic analyses.
-
-In the following applications, advanced computer experiments are performed on the FMUs.
-
-.. note::
-   Here we demonstrate the potential of using OpenTURNS on FMUs.
-   For easier examples, see the section :doc:`Examples<../example/index>`.
-
-.. toctree::
-   :maxdepth: 1
-
-   ../auto_application/index.rst
