+++ doc/examples/doc_FMUPointToFieldFunction_example/GALLERY_HEADER.rst
@@ -1,0 +1,8 @@
+:hide-sidebar-primary: true
+
+Run simulations with time-independent inputs
+============================================
+
+If your model's inputs are time-independent scalars, you can handle it with an object `FMUPointToFieldFunction`.
+These examples below show you how to proceed.
+
