+++ doc/examples/doc_FMUPointToFieldFunction_example/subgallery_beam/GALLERY_HEADER.rst
@@ -1,0 +1,8 @@
+The cantilever beam
+-------------------
+
+These two examples rely on a very simple time-independent model to understand
+how to use OTFMI.
+The first one show you how to use the `FMUPointToFieldFunction` object to run
+simulations, while the second describes how to run simulations
+with OTFMI's low-level functions.
