+++ doc/examples/doc_FMUPointToFieldFunction_example/subgallery_epid/GALLERY_HEADER.rst
@@ -1,0 +1,5 @@
+The epidemiological model
+-------------------------
+
+Here is another set of examples based on model of disease propagation.
+The inputs are still time-independent, but now, the output is time-dependent.
