+++ doc/examples/doc_FieldFunction_example/GALLERY_HEADER.rst
@@ -1,0 +1,7 @@
+:hide-sidebar-primary: true
+
+Run simulations with time-independent inputs
+============================================
+
+If your model's inputs are time-dependent variables, you can handle it with an object `FMUFieldFunction`.
+
