+++ doc/examples/doc_application/GALLERY_HEADER.rst
@@ -1,0 +1,7 @@
+Applications
+============
+
+OTFMI bridges the physics of the model (in `FMI <https://fmi-standard.org/>`__ format) with the universe of
+`OpenTURNS <http://openturns.github.io/openturns/master/contents.html>`__ probabilistic analyses.
+
+In the following applications, advanced computer experiments are performed on the FMUs.
