+++ doc/examples/doc_application_persalys/Solar.mo
