+++ doc/examples/doc_application_persalys/linux64/Solar.fmu
