+++ doc/examples/doc_application_persalys/win64/Solar.fmu
