+++ doc/examples/doc_export_example/GALLERY_HEADER.rst
@@ -1,0 +1,6 @@
+From OpenTURNS to FMI
+---------------------
+
+The focus of OTFMI module is to enable FMUs analysis using OpenTURNS.
+Once a computationnally heavy FMU is metamodeled, the modeler may want to employ
+the metamodel *instead of the FMU* in the simulation tool.
