--- doc/fmus/index.rst
@@ -1,8 +1,0 @@
-Sample FMU files
-================
-
-.. toctree::
-   :maxdepth: 1
-
-   deviation
-   epid
