--- doc/install.rst
@@ -1,6 +1,0 @@
-Installation
-============
-
-.. include:: ../README.rst
-    :start-line: 24
-    :end-line: 29
