+++ doc/sg_execution_times.rst
@@ -1,0 +1,64 @@
+
+:orphan:
+
+.. _sphx_glr_sg_execution_times:
+
+
+Computation times
+=================
+**00:00.015** total execution time for 10 files **from all galleries**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_examples_auto_FieldFunction_example_ex_Field.py` (``examples/doc_FieldFunction_example/ex_Field.py``)
+     - 00:00.015
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_FMUPointToFieldFunction_example_subgallery_beam_ex_PointToField.py` (``examples/doc_FMUPointToFieldFunction_example/subgallery_beam/ex_PointToField.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_FMUPointToFieldFunction_example_subgallery_beam_ex_pyfmi_lowlevel.py` (``examples/doc_FMUPointToFieldFunction_example/subgallery_beam/ex_pyfmi_lowlevel.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_FMUPointToFieldFunction_example_subgallery_epid_ex_dyn_basics.py` (``examples/doc_FMUPointToFieldFunction_example/subgallery_epid/ex_dyn_basics.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_FMUPointToFieldFunction_example_subgallery_epid_ex_dyn_init.py` (``examples/doc_FMUPointToFieldFunction_example/subgallery_epid/ex_dyn_init.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_FMUPointToFieldFunction_example_subgallery_epid_ex_dyn_set.py` (``examples/doc_FMUPointToFieldFunction_example/subgallery_epid/ex_dyn_set.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_application_ex_metamodel.py` (``examples/doc_application/ex_metamodel.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_application_ex_threshold.py` (``examples/doc_application/ex_threshold.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_export_example_ex_fmu_exporter.py` (``examples/doc_export_example/ex_fmu_exporter.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_examples_auto_export_example_ex_model_exporter.py` (``examples/doc_export_example/ex_model_exporter.py``)
+     - 00:00.000
+     - 0.0
